Auth included in flujo caja

+++ src/controllers/flujo_caja.ts
@@ -15,7 +15,12 @@ export const get_flujo_caja = async ( req: Request, res: Response, next: NextFun
         if ( !flujo_caja ) {
             return next( createError( 'Flujo caja not found', 404 ) );
         }
-        res.status( 200 ).json( flujo_caja );
+
+        if( req.user.isAdmin || req.user.empresas.includes( +flujo_caja.empresa_id ) ){
+            res.status( 200 ).json( flujo_caja );
+        }else{
+            next( createError( 'Unauthorized', 401 ) );
+        }
     }catch(error){
         next(error)
     }
@@ -57,6 +62,10 @@ export const create_flujo_caja = async ( req: Request, res: Response, next: Next
             return next( createError( 'Empresa not found', 404 ) );
         }
 
+        if(!req.user.isAdmin && !req.user.empresas.includes( +empresa_id )){
+            return next( createError( 'Unauthorized', 401 ) );
+        }
+
         const flujo_caja = await prisma.flujo_caja.create({
             data: {
                 fecha: new Date(fecha),
@@ -79,6 +88,13 @@ export const update_flujo_caja = async ( req: Request, res: Response, next: Next
             saldo_anterior
         } = req.body;
 
+        const flujo_caja = await prisma.flujo_caja.findUnique({
+            where: { id: +id },
+        });
+        if (!req.user.isAdmin && !req.user.empresas.includes( +flujo_caja.empresa_id )) {
+            return next( createError( 'Unauthorized', 401 ) );
+        }
+
         if(empresa_id){
             const empresa = await prisma.empresa.findUnique({
                 where: { id: +empresa_id },
@@ -87,6 +103,10 @@ export const update_flujo_caja = async ( req: Request, res: Response, next: Next
             if(!empresa){
                 return next( createError( 'Empresa not found', 404 ) );
             }
+
+            if(!req.user.isAdmin && !req.user.empresas.includes( +empresa_id )){
+                return next( createError( 'Unauthorized', 401 ) );
+            }
         }
 
         let new_fecha = undefined
@@ -94,7 +114,7 @@ export const update_flujo_caja = async ( req: Request, res: Response, next: Next
             new_fecha = new Date(fecha);
         }
 
-        const flujo_caja = await prisma.flujo_caja.update({
+        const flujo_caja_updated = await prisma.flujo_caja.update({
             where: { id: +id },
             data: {
                 fecha: new_fecha,
@@ -102,7 +122,7 @@ export const update_flujo_caja = async ( req: Request, res: Response, next: Next
                 saldo_anterior: saldo_anterior || undefined
             },
         });
-        res.status( 200 ).json( flujo_caja );
+        res.status( 200 ).json( flujo_caja_updated );
     }catch(error){
         next(error);
     }
@@ -111,10 +131,17 @@ export const update_flujo_caja = async ( req: Request, res: Response, next: Next
 export const delete_flujo_caja = async ( req: Request, res: Response, next: NextFunction ) => {
     try {
         const { id } = req.params;
-        const flujo_caja = await prisma.flujo_caja.delete({
+        const flujo_caja = await prisma.flujo_caja.findUnique({
             where: { id: +id },
         });
-        res.status( 200 ).json( flujo_caja );
+        if (!req.user.isAdmin && !req.user.empresas.includes( +flujo_caja.empresa_id )) {
+            return next( createError( 'Unauthorized', 401 ) );
+        }
+
+        const flujo_caja_deleted = await prisma.flujo_caja.delete({
+            where: { id: +id },
+        });
+        res.status( 200 ).json( flujo_caja_deleted );
     }catch(error){
         next(error);
     }
@@ -131,7 +158,16 @@ export const get_categoria = async ( req: Request, res: Response, next: NextFunc
         if ( !categoria ) {
             return next( createError( 'Categoria not found', 404 ) );
         }
-        res.status( 200 ).json( categoria );
+
+        //validate auth
+        const flujo_caja = await prisma.flujo_caja.findUnique({
+            where: { id: +categoria.flujo_caja_id },
+        });
+        if (req.user.isAdmin || req.user.empresas.includes( +flujo_caja.empresa_id )) {
+            res.status( 200 ).json( categoria );
+        }else{
+            next( createError( 'Unauthorized', 401 ) );
+        }
     }catch(error){
         next(error)
     }
@@ -143,8 +179,14 @@ export const get_categorias_by_flujo_caja = async ( req: Request, res: Response,
         const flujo_caja = await prisma.flujo_caja.findUnique({
             where: { id: +flujo_caja_id },
         });
+        
         if ( !flujo_caja ) {
             return next( createError( 'Flujo caja not found', 404 ) );
+        }
+
+        //validate auth
+        if (!req.user.isAdmin && !req.user.empresas.includes( +flujo_caja.empresa_id )) {
+            return next( createError( 'Unauthorized', 401 ) );
         }
 
         const categorias = await prisma.categoria.findMany({
@@ -175,6 +217,10 @@ export const create_categoria = async ( req: Request, res: Response, next: NextF
             return next( createError( 'Flujo caja not found', 404 ) );
         }
 
+        if(!req.user.isAdmin && !req.user.empresas.includes( +flujo_caja.empresa_id )){
+            return next( createError( 'Unauthorized', 401 ) );
+        }
+
         const categoria = await prisma.categoria.create({
             data: {
                 descripcion: descripcion,
@@ -203,6 +249,20 @@ export const update_categoria = async ( req: Request, res: Response, next: NextF
             flujo_caja_id
         } = req.body;
 
+        //validate auth
+        const categoria = await prisma.categoria.findUnique({
+            where: { id: +id },
+        });
+
+        const flujo_caja = await prisma.flujo_caja.findUnique({
+            where: { id: +categoria.flujo_caja_id },
+        });
+
+        if(!req.user.isAdmin && !req.user.empresas.includes( +flujo_caja.empresa_id )){
+            return next( createError( 'Unauthorized', 401 ) );
+        }
+
+        //Validates new flujo_caja_id
         if(flujo_caja_id){
             const flujo_caja = await prisma.flujo_caja.findUnique({
                 where: { id: +flujo_caja_id },
@@ -211,9 +271,13 @@ export const update_categoria = async ( req: Request, res: Response, next: NextF
             if(!flujo_caja){
                 return next( createError( 'Flujo caja not found', 404 ) );
             }
+
+            if(!req.user.isAdmin && !req.user.empresas.includes( +flujo_caja.empresa_id )){
+                return next( createError( 'Unauthorized', 401 ) );
+            }
         }
 
-        const categoria = await prisma.categoria.update({
+        const categoria_updated = await prisma.categoria.update({
             where: { id: +id },
             data: {
                 descripcion: descripcion || undefined,
@@ -224,7 +288,7 @@ export const update_categoria = async ( req: Request, res: Response, next: NextF
                 flujo_caja_id: flujo_caja_id || undefined
             },
         });
-        res.status( 200 ).json( categoria );
+        res.status( 200 ).json( categoria_updated );
     }catch(error){
         next(error);
     }
@@ -233,10 +297,24 @@ export const update_categoria = async ( req: Request, res: Response, next: NextF
 export const delete_categoria = async ( req: Request, res: Response, next: NextFunction ) => {
     try {
         const { id } = req.params;
-        const categoria = await prisma.categoria.delete({
+
+        //validate auth
+        const categoria = await prisma.categoria.findUnique({
             where: { id: +id },
         });
-        res.status( 200 ).json( categoria );
+
+        const flujo_caja = await prisma.flujo_caja.findUnique({
+            where: { id: +categoria.flujo_caja_id },
+        });
+
+        if(!req.user.isAdmin && !req.user.empresas.includes( +flujo_caja.empresa_id )){
+            return next( createError( 'Unauthorized', 401 ) );
+        }
+
+        const categoria_deleted = await prisma.categoria.delete({
+            where: { id: +id },
+        });
+        res.status( 200 ).json( categoria_deleted );
     }catch(error){
         next(error);
     }
